requirements.txt
matplotlib==3.8.1
numpy==1.26.1
opencv-python==4.8.1.78
python-telegram-bot==21.4
future~=0.18.3
pillow~=10.3.0
requests~=2.31.0
tk==0.1.0